+++ docs/release-notes.rst
@@ -1,6 +1,7 @@
 .. This work is licensed under a Creative Commons Attribution 4.0 International License.
 .. http://creativecommons.org/licenses/by/4.0
-.. Copyright (C) 2021 Nordix
+.. Copyright (C) 2021-2023 Nordix. All rights reserved.
+.. Copyright (C) 2023 OpenInfra Foundation Europe. All rights reserved.
 
 =============
 Release-Notes
@@ -18,7 +19,7 @@ Functions, previously here, but now with their own repos have their own release 
 * :doc:`Information Coordinator Service <informationcoordinatorservice:release-notes>`.
 * :doc:`A1 Policy Management Service <a1policymanagementservice:release-notes>`.
 * :doc:`Near-RT RIC A1 Simulator <simulator:release-notes>`.
-* DMaaP/Kafka Information Producer Adapters. :doc:`DMaaPAdapter <dmaapadapter:release-notes>`. :doc:`DMaaP Mediator Producer <dmaapmediatorproducer:release-notes>`.
+* DMaaP/Kafka Information Producer Adapters. :doc:`DMaaP Adapter <dmaapadapter:release-notes>`. :doc:`DMaaP Mediator Producer <dmaapmediatorproducer:release-notes>`.
 * :doc:`Non-RT-RIC App Catalogue <rappcatalogue:release-notes>`.
 * :doc:`K8S Helm Chart LCM Manager <helmmanager:release-notes>`.
 * :doc:`Service Management & Exposure (SME) <sme:release-notes>`.
@@ -28,6 +29,9 @@ Functions, previously here, but now with their own repos have their own release 
   * :doc:`"Helloworld" O-DU Slice Assurance use case <ransliceassurance:release-notes>`.
 
 
+
+Release-Notes for just the nonrtric repo
+========================================
 
 Bronze
 ------
@@ -198,6 +202,20 @@ H Release
 +-----------------------------+---------------------------------------------------+
 | **Purpose of the delivery** | nonrtric-plt-auth-token-fetch:1.1.1               |
 |                             | Updated Springboot version                        |
+|                             |                                                   |
++-----------------------------+---------------------------------------------------+
+
+I Release
+---------
++-----------------------------+---------------------------------------------------+
+| **Project**                 | Non-RT RIC                                        |
+|                             |                                                   |
++-----------------------------+---------------------------------------------------+
+| **Repo/commit-ID**          | nonrtric                                          |
+|                             |                                                   |
++-----------------------------+---------------------------------------------------+
+| **Note**                    | No new images released from this repo for         |
+|                             |     the I Release.                                |
 |                             |                                                   |
 +-----------------------------+---------------------------------------------------+